Change load cell value from 32-bit int to 16-bit int
It should not be a 32-bit int with the calibration factor changes, and the read() method returns a 16-bit int anyway

+++ lib/LoadCell/LoadCell.h
@@ -17,7 +17,7 @@ private:
     int8_t sck;
     int8_t sda;
     int32_t calibration;
-    int32_t value;
+    int16_t value;
     HX711 lc;
 };
